requirements.txt
pywinauto
uiautomation
psutil
pyautogui
pyperclip
Pillow
whatsapp-python
groq
google-genai
streamlit
SpeechRecognition
pyaudio
pyttsx3
opencv-python
PyYAML
python-dotenv
watchdog
customtkinter
packaging
